requirements.txt
mcp[cli]
python-pptx
aspose.slides
cairosvg
pytest
lxml
